emotion-main/requirements.txt
absl-py==2.4.0
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
async-timeout==5.0.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
decorator==4.4.2
exceptiongroup==1.3.1
facenet-pytorch==2.6.0
fastapi==0.116.1
ffmpeg-python==0.2.0
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
Flask-SSE==1.0.0
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.3.0
future==1.0.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
hsemotion-onnx==0.3.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.12.1
kiwisolver==1.5.0
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.33
ml_dtypes==0.5.4
moviepy==1.0.3
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
numpy==1.26.4
onnx==1.21.0
onnxruntime==1.16.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.19.0
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pandas==2.3.3
pillow==10.2.0
proglog==0.1.12
prometheus_client==0.21.1
protobuf==7.34.1
psutil==7.2.2
py-cpuinfo==9.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.2
redis==7.4.0
requests==2.33.0
rich==14.3.3
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sounddevice==0.5.5
starlette==0.47.3
sympy==1.14.0
tensorflow==2.21.0
termcolor==3.3.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.72
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.35.0
Werkzeug==3.1.7
wrapt==2.1.2
